Create 4_Olya_and_Game_with_Arrays.cpp
B. Olya and Game with Arrays
time limit per test1 second
memory limit per test256 megabytes
Artem suggested a game to the girl Olya. There is a list of n
 arrays, where the i
-th array contains mi≥2
 positive integers ai,1,ai,2,…,ai,mi
.

Olya can move at most one (possibly 0
) integer from each array to another array. Note that integers can be moved from one array only once, but integers can be added to one array multiple times, and all the movements are done at the same time.

The beauty of the list of arrays is defined as the sum ∑ni=1minmij=1ai,j
. In other words, for each array, we find the minimum value in it and then sum up these values.

The goal of the game is to maximize the beauty of the list of arrays. Help Olya win this challenging game!

Input
Each test consists of multiple test cases. The first line contains a single integer t
 (1≤t≤25000
) — the number of test cases. The description of test cases follows.

The first line of each test case contains a single integer n
 (1≤n≤25000
) — the number of arrays in the list.

This is followed by descriptions of the arrays. Each array description consists of two lines.

The first line contains a single integer mi
 (2≤mi≤50000
) — the number of elements in the i
-th array.

The next line contains mi
 integers ai,1,ai,2,…,ai,mi
 (1≤ai,j≤109
) — the elements of the i
-th array.

It is guaranteed that the sum of mi
 over all test cases does not exceed 50000
.

Output
For each test case, output a single line containing a single integer — the maximum beauty of the list of arrays that Olya can achieve.

Example
InputCopy
3
2
2
1 2
2
4 3
1
3
100 1 6
3
4
1001 7 1007 5
3
8 11 6
2
2 9
OutputCopy
5
1
19
Note
In the first test case, we can move the integer 3
 from the second array to the first array. Then the beauty is min(1,2,3)+min(4)=5
. It can be shown that this is the maximum possible beauty.

In the second test case, there is only one array, so regardless of the movements, the beauty will be min(100,1,6)=1
.